chore: reorganize constants, no more index
