+++ packages/bridge-controller/src/constants/chains.ts
@@ -1,3 +1,5 @@
+import type { AllowedBridgeChainIds } from './bridge';
+
 /**
  * An object containing all of the chain ids for networks both built in and
  * those that we have added custom code to support our feature set.
@@ -155,3 +157,17 @@ export const NETWORK_TO_NAME_MAP = {
   [CHAIN_IDS.LISK]: LISK_DISPLAY_NAME,
   [CHAIN_IDS.LISK_SEPOLIA]: LISK_SEPOLIA_DISPLAY_NAME,
 } as const;
+export const NETWORK_TO_SHORT_NETWORK_NAME_MAP: Record<
+  AllowedBridgeChainIds,
+  string
+> = {
+  [CHAIN_IDS.MAINNET]: 'Ethereum',
+  [CHAIN_IDS.LINEA_MAINNET]: 'Linea',
+  [CHAIN_IDS.POLYGON]: NETWORK_TO_NAME_MAP[CHAIN_IDS.POLYGON],
+  [CHAIN_IDS.AVALANCHE]: 'Avalanche',
+  [CHAIN_IDS.BSC]: NETWORK_TO_NAME_MAP[CHAIN_IDS.BSC],
+  [CHAIN_IDS.ARBITRUM]: NETWORK_TO_NAME_MAP[CHAIN_IDS.ARBITRUM],
+  [CHAIN_IDS.OPTIMISM]: NETWORK_TO_NAME_MAP[CHAIN_IDS.OPTIMISM],
+  [CHAIN_IDS.ZKSYNC_ERA]: 'ZkSync Era',
+  [CHAIN_IDS.BASE]: 'Base',
+};
