+++ packages/bridge-controller/src/types.ts
@@ -12,7 +12,7 @@ import type { Hex } from '@metamask/utils';
 import type { BigNumber } from 'bignumber.js';
 
 import type { BridgeController } from './bridge-controller';
-import type { BRIDGE_CONTROLLER_NAME } from './constants';
+import type { BRIDGE_CONTROLLER_NAME } from './constants/bridge';
 
 /**
  * The types of assets that a user can send
