+++ packages/bridge-controller/src/utils/index.test.ts
@@ -11,11 +11,14 @@ import {
   sumHexes,
   getBridgeApiBaseUrl,
 } from '.';
-import { ETH_USDT_ADDRESS, METABRIDGE_ETHEREUM_ADDRESS } from '../constants';
+import {
+  ETH_USDT_ADDRESS,
+  METABRIDGE_ETHEREUM_ADDRESS,
+} from '../constants/bridge';
 import {
   BRIDGE_DEV_API_BASE_URL,
   BRIDGE_PROD_API_BASE_URL,
-} from '../constants';
+} from '../constants/bridge';
 import { CHAIN_IDS } from '../constants/chains';
 import { SWAPS_CHAINID_DEFAULT_TOKEN_MAP } from '../constants/tokens';
