+++ packages/bridge-controller/src/utils/index.ts
@@ -3,12 +3,12 @@ import type { Hex } from '@metamask/utils';
 import { Contract } from 'ethers';
 
 import {
+  DEFAULT_BRIDGE_CONTROLLER_STATE,
   BRIDGE_DEV_API_BASE_URL,
   BRIDGE_PROD_API_BASE_URL,
-  DEFAULT_BRIDGE_CONTROLLER_STATE,
   ETH_USDT_ADDRESS,
   METABRIDGE_ETHEREUM_ADDRESS,
-} from '../constants';
+} from '../constants/bridge';
 import { CHAIN_IDS } from '../constants/chains';
 import { SWAPS_CHAINID_DEFAULT_TOKEN_MAP } from '../constants/tokens';
 import type { BridgeControllerState } from '../types';
